agents/orchestrator/requirements.txt
httpx==0.25.1
hyperliquid-python-sdk>=0.4.0
